Send Config Info working successfully

+++ BackEnd/BallSpinner/PreviousThrow.cs
@@ -200,7 +200,7 @@ public class PreviousThrow : IBallSpinner
     }
 
     /// <inheritdoc/>
-    public void SubmitSmartDotConfig(int[] ODRs, int[] SampleRates)
+    public void SubmitSmartDotConfig(double[] ODRs, double[] SampleRates)
     {
         if (!IsSmartDotPaired())
             return;
